Separate integration tests away into a separate folder. (#324)
* Separate tests that can be run locally from tests that needs to connect to network (i.e. integration tests).

* Fix relative path issues in the integration test folder.
