+++ Makefile
@@ -1,4 +1,4 @@
 .PHONY: test
 
 test:
-	nosetests -d -v -A 'stable' test/test*.py
+	nosetests -d -v test/test*.py
