+++ test/integration/test_binance_order_book_tracker.py
@@ -2,7 +2,7 @@
 
 from os.path import join, realpath
 import sys
-sys.path.insert(0, realpath(join(__file__, "../../")))
+sys.path.insert(0, realpath(join(__file__, "../../../")))
 
 from hummingbot.market.binance.binance_order_book_tracker import BinanceOrderBookTracker
 import asyncio
