+++ test/integration/test_coinbase_pro_order_book_tracker.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 from os.path import join, realpath
-import sys; sys.path.insert(0, realpath(join(__file__, "../../")))
+import sys; sys.path.insert(0, realpath(join(__file__, "../../../")))
 
 import asyncio
 import logging
