+++ test/integration/test_ddex_market.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 from os.path import join, realpath
-import sys; sys.path.insert(0, realpath(join(__file__, "../../")))
+import sys; sys.path.insert(0, realpath(join(__file__, "../../../")))
 
 import asyncio
 import conf
