+++ test/integration/test_order_expiration.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 from os.path import join, realpath
 import sys
-sys.path.insert(0, realpath(join(__file__, "../../")))
+sys.path.insert(0, realpath(join(__file__, "../../../")))
 import pandas as pd
 from typing import (
     List,
